moved WF machinery
