+++ tests/__init__.py
@@ -31,8 +31,6 @@ import unittest
 
 import msprime
 
-from .wf import wf_sim
-
 NULL_NODE = -1
