--- tests/wf/__init__.py
@@ -1,2 +1,0 @@
-from .wf import wf_sim
-assert wf_sim  # to silence flake8
